Ensure object-dumper prebackuppod has writeable home directory
After some experimentation with [FlowSchemas], we found that the API
interactions get throttled even when running the PrebackupPod with the
`system` flow priority (same level as the Kubelets).

We then found that the throttling is caused because `kubectl` isn't able
to create its API cache (in `$HOME/.kube/cache` and
`$HOME/.kube/http-cache`) on OpenShift 4, because we don't set `$HOME`
to a writeable location, and the pod runs as an unprivileged user on
OpenShift.

On plain Kubernetes the issue doesn't occur, because the PrebackupPod
runs as root there, and thus can write to `/.kube`.

This commit ensures that `$HOME` is set to `/home/dumper` and that
`/home/dumper` is writeable, by mounting an emptyDir volume there.

Fixes projectsyn/k8s-object-dumper#8.

[FlowSchemas]: https://kubernetes.io/docs/concepts/cluster-administration/flow-control